requirements.txt
SQLAlchemy==2.0.35
alembic==1.13.3
asyncpg==0.29.0
fastapi==0.115.2
uvicorn==0.31.1
pydantic==2.9.2
pydantic_core==2.23.4
pydantic[email]
python-jose==3.3.0
passlib==1.7.4
bcrypt==4.0.0
python-multipart==0.0.12
python-dotenv==1.0.1
websockets==13.1
aiogram==3.13.1
redis==5.1.1
celery==5.4.0
jinja2==3.1.4